requirements.txt
fastapi
uvicorn[standard]
pydantic
redis
python-multipart
transformers[torch]
librosa
numpy
webrtcvad-wheels
soundfile
pytest
locust
prometheus-client
